backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
python-dotenv
openai>=1.40.0
tiktoken>=0.7.0
